SpecializationList tests and DevelopmentList tests moved to their respective folders
